user tests added

+++ src/main/java/cz/cvut/fit/tjv/semestral/data/entities/User.java
@@ -8,7 +8,7 @@ import java.util.Collection;
 public class User {
     @Id
     @GeneratedValue
-    private Integer id;  // PK
+    private Long id;  // PK
 
     @OneToOne
     private Authentication login;  // FK
@@ -25,11 +25,11 @@ public class User {
     private String displayedName;
     private String email;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
